requirements.txt
# ── Core ─────────────────────────────────────────────
torch==2.11.0
torchvision
numpy==2.4.3
scipy==1.17.1
scikit-learn==1.8.0
Pillow
PyYAML==6.0.3
tqdm==4.67.3
packaging==26.0

# ── Diffusers (InstructPix2Pix / SD-Turbo / SwiftEdit) ──
diffusers
accelerate
safetensors==0.7.0

# ── HuggingFace (CLIP eval & tokenizer) ─────────────
transformers==5.3.0
tokenizers==0.22.2
huggingface_hub==1.7.2
sentence-transformers==5.3.0

# ── Evaluation (FID / IS / CLIP) ────────────────────
pytorch-fid

# ── Data ─────────────────────────────────────────────
lmdb

# ── Misc ─────────────────────────────────────────────
filelock==3.25.2
fsspec==2026.2.0
requests==2.32.5
regex==2026.2.28
Jinja2==3.1.6
sympy==1.14.0
networkx==3.6.1
rich==14.3.3
typer==0.24.1
